Refactor platform stores into modular slices
